requirements.txt
arxiv==1.4.8
BCEmbedding==0.1.3
beautifulsoup4==4.12.3
cachetools==5.3.2
chromadb==0.4.22
fake_useragent==1.4.0
gradio==4.21.0
gradio_modal==0.0.2
grpcio==1.60.1
httpx==0.27.0
huggingface_hub==0.20.2
langchain==0.1.13
langchain_community==0.0.29
langchain_core==0.1.33
langchain_openai==0.1.1
langchain_text_splitters==0.0.1
librosa==0.10.1
openai==1.14.2
pathvalidate==3.2.0
pdfplumber==0.10.4
pydantic==2.6.4
PyJWT==2.8.0
pypdf==4.1.0
python-dotenv==1.0.1
Requests==2.31.0
scholarly==1.7.11
typing_extensions==4.10.0
websockets==10.4
zhipuai==2.0.1
uvicorn==0.20.0
fastapi==0.103.0
git+https://github.com/wenet-e2e/wenet.git